Enable more compiler warnings for debug version

+++ src/scriptable/scriptable.cpp
@@ -724,7 +724,7 @@ void Scriptable::action()
                                                   : QString('\n');
         m_proxy->action(data, command);
     } else {
-        QByteArray message = QByteArray::number((qlonglong)m_proxy->openActionDialog(data));
+        QByteArray message = QByteArray::number(m_proxy->openActionDialog(data));
         sendMessageToClient(message, CommandActivateWindow);
     }
 }
